console: fix json store query editor
