+++ console2/src/components/molecules/ProjectConfiguration/index.tsx
@@ -19,10 +19,10 @@
  */
 
 import React, { useRef, useState } from 'react';
-import { Button, Loader } from 'semantic-ui-react';
-import Editor from '@monaco-editor/react';
+import { Button } from 'semantic-ui-react';
 import _ from 'lodash';
 import './styles.css';
+import LoadingEditor from '../LoadingEditor';
 
 interface Props {
     config?: Object;
@@ -59,8 +59,9 @@ const ProjectConfiguration: React.FunctionComponent<Props> = ({ config, submitti
 
     const loading = submitting || !isEditorReady;
     return (
-        <div className={'editorContainer'}>
+        <div className={'projectEditorContainer'}>
             <LoadingEditor
+                language="json"
                 initValue={JSON.stringify(config, null, 4)}
                 handleEditorDidMount={handleEditorDidMount}
                 disabled={loading}
@@ -77,27 +78,6 @@ const ProjectConfiguration: React.FunctionComponent<Props> = ({ config, submitti
                 {jsonError && <p style={{ color: 'red', marginTop: '15px' }}>{jsonError}</p>}
             </div>
         </div>
-    );
-};
-
-interface LoadingEditorProps {
-    handleEditorDidMount: (getEditorValue: () => string) => void;
-    initValue?: string;
-    disabled: boolean;
-}
-
-const LoadingEditor = ({ handleEditorDidMount, initValue, disabled }: LoadingEditorProps) => {
-    if (!initValue) {
-        return <Loader active={true} />;
-    }
-
-    return (
-        <Editor
-            language="json"
-            value={initValue}
-            editorDidMount={handleEditorDidMount}
-            options={{ lineNumbers: 'on', minimap: { enabled: false }, readOnly: disabled }}
-        />
     );
 };
